Remove packed statsd, add config to point to statsd if desired
This was mostly like this when the crossposter was being deployed to
heroku and it was easier to do so. But it introduces dependencies on
node 6 that are not needed for the rest.
