+++ .env.example
@@ -25,3 +25,8 @@ CROSSPOSTER_ADMIN_TWITTER=crossposter_admin
 # Fediverse account that users can use to reach the admin in case of issues
 CROSSPOSTER_ADMIN_FEDI_ADDRESS=https://instance.example.com/@crossposter_admin
 CROSSPOSTER_ADMIN_FEDI_AT=crossposter_admin@instance.example.com
+
+# If you are using statsd to save metrics, you need the three following variables
+STATSD_ENABLED=true
+STATSD_HOST=127.0.0.1
+STATSD_PORT=8125
