+++ Procfile
@@ -1,2 +1,2 @@
 web: bundle exec foreman start -f Procfile.web
-worker: bundle exec foreman start -f Procfile.worker
+worker: bundle exec ruby script/daemon.rb run check_for_statuses.rb
