--- lib/stats.rb
@@ -1,21 +1,0 @@
-# frozen_string_literal: true
-
-require 'statsd'
-
-class Stats
-  def initialize
-    @statsd = Statsd.new('127.0.0.1', 8125)
-  end
-
-  def increment(name)
-    @statsd.increment("#{name}")
-  end
-
-  def timing(name, ms)
-    @statsd.timing(name, ms)
-  end
-
-  def time(name, &blk)
-    @statsd.time(name, &blk)
-  end
-end
