+++ package.json
@@ -28,8 +28,6 @@
     "rails-erb-loader": "^5.1.0",
     "resolve-url-loader": "^2.1.0",
     "sass-loader": "^6.0.6",
-    "statsd": "^0.8.0",
-    "statsd-librato-backend": "^2.0.14",
     "style-loader": "^0.18.2",
     "webpack": "^3.5.3",
     "webpack-manifest-plugin": "^1.3.0",
@@ -39,6 +37,6 @@
     "webpack-dev-server": "^2.7.1"
   },
   "engines": {
-    "node": "6.x"
+    "node": "8.x"
   }
 }
